feat: add proposal icons to proposal card
