--- src/pages/CurrentEntity/Overview/InstructionsToExecute/index.ts
@@ -1,1 +1,0 @@
-export { default as InstructionsToExecute } from './InstructionsToExecute'
