+++ src/utils/dao.ts
@@ -1,11 +1,18 @@
 import { contracts, customQueries } from '@ixo/impactxclient-sdk'
 import { CosmWasmClient } from '@ixo/impactxclient-sdk/node_modules/@cosmjs/cosmwasm-stargate'
 import { MarketingInfoResponse, TokenInfoResponse } from '@ixo/impactxclient-sdk/types/codegen/Cw20Base.types'
-import { ProposalResponse, Threshold, VoteInfo } from '@ixo/impactxclient-sdk/types/codegen/DaoProposalSingle.types'
+import {
+  ProposalResponse,
+  Threshold,
+  VoteInfo,
+  CosmosMsgForEmpty,
+} from '@ixo/impactxclient-sdk/types/codegen/DaoProposalSingle.types'
 import { UpdateProposalConfigData } from 'components/Modals/AddActionModal/SetupUpdateVotingConfigModal'
 import { chainNetwork } from 'hooks/configs'
 import { Member } from 'types/dao'
 import { durationToSeconds, expirationAtTimeToSecondsFromNow } from './conversions'
+import { ProposalActionConfigMap } from 'types/protocol'
+import { parseEncodedMessage } from './messages'
 
 export const thresholdToTQData = (
   source: Threshold,
@@ -217,4 +224,30 @@ export const getDaoContractMembersInfo = async ({
   }
 
   return members
+}
+
+export const proposalMsgToActionConfig = (msg: CosmosMsgForEmpty) => {
+  if ('wasm' in msg && 'execute' in msg.wasm && 'msg' in msg.wasm.execute) {
+    const encodedMessage = parseEncodedMessage(msg.wasm.execute.msg)
+
+    let key: string = Object.keys(encodedMessage)[0]
+    const value: any = Object.values(encodedMessage)[0]
+
+    if (key === 'update_config') {
+      if ('config' in value) {
+        key += '.config'
+      } else if ('deposit_info' in value) {
+        key += '.proposal'
+      } else if ('threshold' in value) {
+        key += '.voting'
+      }
+    }
+
+    const proposalActionDetail = ProposalActionConfigMap[`wasm.execute.${key}`]
+    return {
+      ...proposalActionDetail,
+      data: value,
+      type: `wasm.execute.${key}`,
+    }
+  }
 }
